requirements.txt
praw
openai
yagmail
python-dotenv